+++ apps/order/serializers.py
@@ -1,17 +1,7 @@
-from .models import Order, Product, Category
+from .models import Order
 from rest_framework import serializers
 
 class OrderSerializer(serializers.ModelSerializer):
     class Meta:
         model = Order
-        fields = "__all__"
-
-class ProductSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Product
-        fields = "__all__"
-
-class CategorySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Category
         fields = "__all__"
