+++ apps/order/views.py
@@ -1,15 +1,7 @@
 from rest_framework import viewsets
-from .serializers import ProductSerializer, OrderSerializer, CategorySerializer
-from .models import Order, Product, Category
+from .serializers import OrderSerializer
+from .models import Order
 
 class OrderViewSet(viewsets.ModelViewSet):
     queryset = Order.objects.all()
     serializer_class = OrderSerializer
-
-class ProductViewSet(viewsets.ModelViewSet):
-    queryset = Product.objects.all()
-    serializer_class = ProductSerializer
-
-class CategoryViewSet(viewsets.ModelViewSet):
-    queryset = Category.objects.all()
-    serializer_class = CategorySerializer
