+++ apps/product/serializers.py
@@ -1,0 +1,8 @@
+from .models import Product
+from rest_framework import serializers
+
+
+class ProductSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Product
+        fields = "__all__"
