+++ apps/product/urls.py
@@ -3,6 +3,7 @@ from . import views
 from rest_framework import routers
 
 router = routers.DefaultRouter()
+router.register(r'product', views.ProductViewSet)
 
 urlpatterns =[
     path('', include(router.urls)),
