+++ apps/product/views.py
@@ -1,0 +1,7 @@
+from rest_framework import viewsets
+from .serializers import ProductSerializer
+from .models import Product
+
+class ProductViewSet(viewsets.ModelViewSet):
+    queryset = Product.objects.all()
+    serializer_class = ProductSerializer
