+++ apps/urls.py
@@ -4,6 +4,7 @@ from django.urls import path, include
 urlpatterns = [
     path('users/', include('apps.user.urls')),
     path('order/', include('apps.order.urls')),
+    path('product/', include('apps.product.urls')),
     # path('schema', SpectacularAPIView.as_view(), name='schema'),
     # path('swagger', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger-ui'),
     # path('redoc', SpectacularRedocView.as_view(url_name='schema'), name='redoc'),
